refactor: extract session and routing logic
